backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
cryptography==44.0.0
PyNaCl==1.5.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==14.1
bittensor==10.0.0
